+++ docs/changelog.md
@@ -5,7 +5,7 @@ All notable changes to OpenBoost will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.0] - 2026-01-20
+## [1.0.0rc1] - 2026-01-20
 
 ### Added